+++ lib/src/parser/character/none_of.dart
@@ -1,23 +1,16 @@
 import 'package:meta/meta.dart';
 
 import '../../core/parser.dart';
-import '../predicate/single_character.dart';
+import '../predicate/character.dart';
 import 'internal/code.dart';
 import 'internal/not.dart';
 import 'internal/optimize.dart';
 
-/// Returns a parser that accepts none of the specified characters (UTF-16 code
-/// units).
+/// Returns a parser that accepts none of the specified characters.
 @useResult
-Parser<String> noneOf(String chars, [String? message]) => SingleCharacterParser(
-    NotCharacterPredicate(optimizedString(chars)),
-    message ?? 'none of "${toReadableString(chars)}" expected');
-
-/// Returns a parser that accepts none of the specified characters (Unicode
-/// code-points).
-@useResult
-Parser<String> noneOfUnicode(String chars, [String? message]) =>
-    SingleCharacterParser(
-        NotCharacterPredicate(optimizedString(chars, unicode: true)),
+Parser<String> noneOf(String chars, {String? message, bool unicode = false}) =>
+    CharacterParser(
+        NotCharacterPredicate(optimizedString(chars, unicode: unicode)),
         message ??
-            'none of "${toReadableString(chars, unicode: true)}" expected');
+            'none of "${toReadableString(chars, unicode: unicode)}" expected',
+        unicode: unicode);